chore: rename mapreceptor2 to mapreceptor
